requirements.txt
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.5.3.56
numpy>=1.19.5
Pillow>=8.2.0
boto3>=1.26.0
requests>=2.26.0
python-dotenv>=0.19.0
ultralytics>=8.0.0  # For YOLOv8
tqdm>=4.62.3